requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
python-dotenv>=1.0.0
pytest>=8.0.0
responses>=0.25.0
ruff>=0.4.0
